Add Vega test data and move main function into test
