+++ src/main/java/org/scijava/plot/specification/VegaPlot.java
@@ -17,14 +17,6 @@ package org.scijava.plot.specification;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import org.scijava.plot.specification.encoding.EncodingChannel;
 import org.scijava.plot.specification.encoding.VegaEncoding;
 
 /**
@@ -188,75 +180,6 @@ public class VegaPlot {
 
     public void setEncoding(VegaEncoding encoding) {
         this.encoding = encoding;
-    }
-
-    public static void main(String args[]) throws IOException {
-
-        VegaPlot plot = new VegaPlot();
-
-        plot.setDescription("A simple bar chart with embedded data.");
-
-        plot.getMark().setType("bar");
-
-        EncodingChannel x = new EncodingChannel();
-        x.setField("a");
-        x.setType("ordinal");
-        plot.getEncoding().setX(x);
-
-        EncodingChannel y = new EncodingChannel();
-        y.setField("b");
-        y.setType("quantitative");
-        plot.getEncoding().setY(y);
-
-        List<Map<String, Object>> data = new ArrayList<>();
-        Map<String, Object> datum;
-
-        datum = new HashMap<>();
-        datum.put("a", "A");
-        datum.put("b", 28);
-        data.add(datum);
-        datum = new HashMap<>();
-        datum.put("a", "B");
-        datum.put("b", 55);
-        data.add(datum);
-        datum = new HashMap<>();
-        datum.put("a", "C");
-        datum.put("b", 43);
-        data.add(datum);
-        datum = new HashMap<>();
-        datum.put("a", "D");
-        datum.put("b", 91);
-        data.add(datum);
-        datum = new HashMap<>();
-        datum.put("a", "E");
-        datum.put("b", 81);
-        data.add(datum);
-        datum = new HashMap<>();
-        datum.put("a", "F");
-        datum.put("b", 53);
-        data.add(datum);
-        datum = new HashMap<>();
-        datum.put("a", "G");
-        datum.put("b", 19);
-        data.add(datum);
-        datum = new HashMap<>();
-        datum.put("a", "H");
-        datum.put("b", 87);
-        data.add(datum);
-        datum = new HashMap<>();
-        datum.put("a", "I");
-        datum.put("b", 52);
-        data.add(datum);
-
-        plot.getData().setValues(data);
-
-        ObjectMapper mapper = new ObjectMapper();
-        mapper.writeValue(new File("/home/hadim/outputfile.json"), plot);
-
-        plot = mapper.readValue(new File("/home/hadim/outputfile.json"), VegaPlot.class);
-
-        System.out.println(mapper.writeValueAsString(plot));
-
     }
 
 }
